2017-10-25 22:35 UTC-0300 Fernando Yurisich <fyurisich@oohg.org>
* core\include\i_altsyntax.ch
* Id.
+ Clause EDITCELLVALUE to grid, browse and xbrowse.
* core\include\i_browse.ch
* Id.
+ Clause EDITCELLVALUE.
* Changed translations of commands SET BROWSESYNC,
SET BROWSEFIXEDBLOCKS and SET BROWSEFIXEDCONTROLS.
* core\include\i_grid.ch
* Id.
+ Clause EDITCELLVALUE.
* core\include\i_pseudofunc.ch
* Id.
+ Added xtranslates for funcions SetBrowseFixedBlocks,
SetBrowseFixedControls, SetBrowseSync, SetXBrowseFixedBlocks
and SetXBrowseFixedControls.
* core\include\i_var.ch
* Id.
+ Added xtranslates for pseudo vars _OOHG_BrowseSyncStatus,
_OOHG_BrowseFixedBlocks, _OOHG_BrowseFixedControls,
_OOHG_XBrowseFixedBlocks and _OOHG_XBrowseFixedControls.
* core\include\i_xbrowse.ch
* Id.
+ Clause EDITCELLVALUE.
* Changed translations of commands SET XBROWSEFIXEDBLOCKS and
SET XBROWSEFIXEDCONTROLS.
* core\source\h_application.prg
* Id.
* Added support for _OOHG_BrowseSyncStatus,
_OOHG_BrowseFixedBlocks, _OOHG_BrowseFixedControls,
_OOHG_XBrowseFixedBlocks and _OOHG_XBrowseFixedControls.
* Changed case style of some words according Harbour's syntax.
* core\source\h_browse.prg
* Id.
- STATIC vars and functions SetBrowseSync, SetBrowseFixedBlocks
and SetBrowseFixedBlocks were substituted by TApplication methods
and pseudo-functions. This should make the class thread safe.
* Changed some comments.
+ Support for clause EDITCELLVALUE bCodeBlock.
The bCodeblock is evaluated by method EditCell2 just
before the cell edition starts.
The codeblock receives 3 parametes: nCol, nRow and uOldValue,
where nRow and nCol identify the cell that will be edited and
uOldValue is the 'face value' to show.
It should return uOldValue or another value of the same type.
Note that changes to uOldValue do not affect the data source.
* Changed Empty( <var> ) to HB_SYMBOL_UNUSED( <var> ).
+ Default value for CheckBeforeEdit parameter at method Define3
of class TOBrowseByCell.
! Method EditGrid of TOBrowseByCell when edition was ended
because of mouse exit.
% Method SetValue of class TOBrowseByCell.
* Methods Left and Right of TOBrowseByCell now return .T. if
and only if row and or col where changed.
* Changed some var names to improve readability.
* core\source\h_grid.prg
* Id.
+ Support for clause EDITCELLVALUE bCodeBlock.
The bCodeblock is evaluated by method EditCell2 just
before the cell edition starts.
The codeblock receives 3 parametes: nCol, nRow and uOldValue,
where nRow and nCol identify the cell that will be edited and
uOldValue is the 'face value' to show.
It should return uOldValue or another value of the same type.
Note that changes to uOldValue do not affect the grid's cell.
* core\source\h_xbrowse.prg
* Id.
- STATIC vars and functions SetXBrowseFixedBlocks and
SetX BrowseFixedBlocks were substituted by TApplication methods
and pseudo-functions. This should make the class thread safe.
+ Support for clause EDITCELLVALUE bCodeBlock.
The bCodeblock is evaluated by method EditCell2 just
before the cell edition starts.
The codeblock receives 3 parametes: nCol, nRow and uOldValue,
where nRow and nCol identify the cell that will be edited and
uOldValue is the 'face value' to show.
It should return uOldValue or another value of the same type.
Note that changes to uOldValue do not affect the data source.
* Some format.